[FLINK-19653][hive] Reduce our dependency on hive runner for tests
