+++ flink-connectors/flink-connector-hive/src/test/java/org/apache/flink/table/catalog/hive/HiveCatalogUseBlinkITCase.java
@@ -20,7 +20,6 @@ package org.apache.flink.table.catalog.hive;
 
 import org.apache.flink.api.common.functions.MapFunction;
 import org.apache.flink.api.java.tuple.Tuple2;
-import org.apache.flink.connectors.hive.FlinkStandaloneHiveRunner;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.streaming.api.functions.sink.SinkFunction;
 import org.apache.flink.table.api.DataTypes;
@@ -48,9 +47,6 @@ import org.apache.flink.types.Row;
 import org.apache.flink.util.CollectionUtil;
 import org.apache.flink.util.FileUtils;
 
-import com.klarna.hiverunner.HiveShell;
-import com.klarna.hiverunner.annotations.HiveSQL;
-import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.ql.udf.UDFMonth;
 import org.apache.hadoop.hive.ql.udf.UDFYear;
 import org.apache.hadoop.hive.ql.udf.generic.GenericUDAFSum;
@@ -60,7 +56,6 @@ import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
-import org.junit.runner.RunWith;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -80,11 +75,7 @@ import static java.lang.String.format;
  * IT case for HiveCatalog.
  * TODO: move to flink-connector-hive-test end-to-end test module once it's setup
  */
-@RunWith(FlinkStandaloneHiveRunner.class)
 public class HiveCatalogUseBlinkITCase extends AbstractTestBase {
-
-	@HiveSQL(files = {})
-	private static HiveShell hiveShell;
 
 	@Rule
 	public TemporaryFolder tempFolder = new TemporaryFolder();
@@ -96,8 +87,7 @@ public class HiveCatalogUseBlinkITCase extends AbstractTestBase {
 
 	@BeforeClass
 	public static void createCatalog() {
-		HiveConf hiveConf = hiveShell.getHiveConf();
-		hiveCatalog = HiveTestUtils.createHiveCatalog(hiveConf);
+		hiveCatalog = HiveTestUtils.createHiveCatalog();
 		hiveCatalog.open();
 	}
 
@@ -248,7 +238,7 @@ public class HiveCatalogUseBlinkITCase extends AbstractTestBase {
 	}
 
 	@Test
-	public void testTimestampUDF() {
+	public void testTimestampUDF() throws Exception {
 
 		TableEnvironment tableEnv = HiveTestUtils.createTableEnvWithBlinkPlannerBatchMode(SqlDialect.HIVE);
 		tableEnv.registerCatalog(hiveCatalog.getName(), hiveCatalog);
@@ -256,7 +246,7 @@ public class HiveCatalogUseBlinkITCase extends AbstractTestBase {
 		tableEnv.executeSql(String.format("create function myyear as '%s'", UDFYear.class.getName()));
 		tableEnv.executeSql("create table src(ts timestamp)");
 		try {
-			HiveTestUtils.createTextTableInserter(hiveShell, "default", "src")
+			HiveTestUtils.createTextTableInserter(hiveCatalog, "default", "src")
 					.addRow(new Object[]{Timestamp.valueOf("2013-07-15 10:00:00")})
 					.addRow(new Object[]{Timestamp.valueOf("2019-05-23 17:32:55")})
 					.commit();
@@ -271,7 +261,7 @@ public class HiveCatalogUseBlinkITCase extends AbstractTestBase {
 	}
 
 	@Test
-	public void testDateUDF() {
+	public void testDateUDF() throws Exception {
 
 		TableEnvironment tableEnv = HiveTestUtils.createTableEnvWithBlinkPlannerBatchMode(SqlDialect.HIVE);
 		tableEnv.registerCatalog(hiveCatalog.getName(), hiveCatalog);
@@ -279,7 +269,7 @@ public class HiveCatalogUseBlinkITCase extends AbstractTestBase {
 		tableEnv.executeSql(String.format("create function mymonth as '%s'", UDFMonth.class.getName()));
 		tableEnv.executeSql("create table src(dt date)");
 		try {
-			HiveTestUtils.createTextTableInserter(hiveShell, "default", "src")
+			HiveTestUtils.createTextTableInserter(hiveCatalog, "default", "src")
 					.addRow(new Object[]{Date.valueOf("2019-01-19")})
 					.addRow(new Object[]{Date.valueOf("2019-03-02")})
 					.commit();
